Back-End/requirements.txt
# Web Automation
selenium>=4.10.0
webdriver-manager>=4.0.0
playwright>=1.56.0

# HTML 解析與數據處理
beautifulsoup4>=4.12.2
lxml>=4.9.3

# HTTP 請求
requests>=2.31.0
aiohttp>=3.9.1

# 日期處理
python-dateutil>=2.8.2
pytz>=2023.3

# Supabase 客戶端
supabase>=2.0.0
postgrest>=0.13.0

# Google Gemini API
google-genai
google-api-core==2.25.0rc1
google-cloud-core==2.3.3

# 定時排程
schedule>=1.2.0
APScheduler>=3.10.4

# 環境變數與設定
python-dotenv>=1.0.0
pydantic>=2.5.1

# 進度顯示與文字處理
tqdm>=4.66.1
unidecode>=1.3.7

# 圖片處理
Pillow>=10.1.0

# Async 支援
nest_asyncio>=1.5.8
asyncio>=3.4.3

# JSON 處理
ujson>=5.8.0

# 型別提示支援
typing-extensions>=4.8.0

# 記錄與監控
loguru>=0.7.2
